Remove oh-my-posh dependency from code
